+++ CHANGES.md
@@ -6,6 +6,7 @@
 * Add TLS support for `io.l5d.httpController` and `io.l5d.mesh`.
 * Add support for HTTP/2 tracing.
 * Cache dtab observations in the io.l5d.consul store.
+* Add Istio namer that uses Istio-Manager's Service Discovery Service.
 
 ## 1.0.2 2017-05-12